refactor(time slider): move time slider to its own component
Moves time slider to its own component (used both in audio controls and video OSD)
